pyproject.toml
[tool.poetry]
name = "kuflow-sdk-python"
version = "0.7.2.dev0"
description = "KuFlow packages to build awesome clients"
license = "MIT"
authors = ["KuFlow S.L. <kuflow@kuflow.com>"]
homepage = "https://kuflow.com/"
documentation = "https://docs.kuflow.com/"
repository = "https://github.com/kuflow/kuflow-sdk-python"
readme = "README.md"
include = ["VERSION", "README.md"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.8"
kuflow-rest = {path="./kuflow-rest", develop=true}
kuflow-robotframework = {path="./kuflow-robotframework", develop=true}
kuflow-temporal-common = {path="./kuflow-temporal-common", develop=true}
kuflow-temporal-activity-kuflow = {path="./kuflow-temporal-activity-kuflow", develop=true}
kuflow-temporal-activity-robotframework = {path="./kuflow-temporal-activity-robotframework", develop=true}

[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
flake8 = "^5.0.4"
black = "^23.3.0"
pytest = "^7.3.1"
pyyaml = "^6.0"
isort = "^5.12.0"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
python_files= "test_*.py"

[tool.black]
line-length = 120
target-version = ["py38", "py39"]

[tool.isort]
profile = "black"
line_length = 120
skip_glob = ["*/_generated/*"]
skip_gitignore = true
known_first_party = ["kuflow_rest", "kuflow_temporal_common", "kuflow_temporal_activity_kuflow"]
